Use urllib2 instead of HTTPConnection
urlib2 is a higher level access for http in python and allows for automatic
proxy control as well as error handling greatly simplyfying the code.  Urllib2
was also added in python 2.1 which should cover anything in the wild.
Hopefully no one is running anything less than 2.5 or 2.6.

Also switched the url to https.  Certificate validation is not turned on until
Python 2.7.9 or 3.4.3 but it should help protect a bit against this attack
vector.